Revert "Issue 25: remove the addthis class"
This reverts commit d72437317cffc6bda60f06bba83d0619f9e264b2.
