+++ addthis.services.yml
@@ -1,4 +1,9 @@
 services:
+  # AddThis service?
+  addthis.addthis:
+    class: Drupal\addthis\AddThis
+    arguments: ['@addthis.script_manager', '@config.factory']
+
   # This service creates a more complex access check.
   addthis.script_manager:
     class: Drupal\addthis\AddThisScriptManager
