+++ src/Plugin/Field/FieldFormatter/AddThisBasicButtonFormatter.php
@@ -10,6 +10,7 @@ use Drupal\addthis\AddThisBasicButtonFormTrait;
 use Drupal\Core\Field\FormatterBase;
 use Drupal\Core\Field\FieldItemListInterface;
 use Drupal\Core\Form\FormStateInterface;
+use Drupal\addthis\AddThis;
 
 
 /**
