+++ src/Tests/AddThisBaseTest.php
@@ -9,9 +9,6 @@ namespace Drupal\addthis\Tests;
 
 use Drupal\simpletest\WebTestBase;
 
-/**
- * Tests the add this functionality.
- */
 class AddThisBaseTest extends WebTestBase {
 
   /**
